webnn/model-tests/image-classification/squeezenet_nhwc.https.html

<!doctype html>
<meta charset=utf-8>
<meta name="timeout" content="long">
<title>test SqueezeNet 1.0 model with NHWC layout</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/libs/numpy.js"></script>
<script src="../resources/utils.js"></script>
<script src="../../resources/utils.js"></script>
<script>
let context;
let builder;
let graph;

const weightsUrl = '../resources/test-data/models/squeezenet1.0_nhwc/weights/';
const outputDimensions = [1, 1001];

const buildConv = async (input, name, options = {}) => {
  const prefix = weightsUrl + name;
  const weightsName = prefix + '_kernel.npy';
  const weights = await buildConstantByNpy(builder, weightsName);
  const biasName = prefix + '_Conv2D_bias.npy';
  const bias = await buildConstantByNpy(builder, biasName);
  options.inputLayout = 'nhwc';
  options.filterLayout = 'ohwi';
  options.bias = bias;
  options.activation = builder.relu();
  return builder.conv2d(input, weights, options);
};

const buildFire = async (input, name) => {
  const convSqueeze = await buildConv(input, name + '_squeeze');
  const convE1x1 = await buildConv(convSqueeze, name + '_e1x1');
  const convE3x3 = await buildConv(
      convSqueeze, name + '_e3x3', {padding: [1, 1, 1, 1]});
  return builder.concat([convE1x1, convE3x3], 3);
};

const load = async () => {
  const strides = [2, 2];
  const layout = 'nhwc';
  const placeholder = builder.input('input',
      {type: 'float32', dimensions: [1, 224, 224, 3]});
  const conv1 = await buildConv(
      placeholder, 'conv1', {strides, autoPad: 'same-upper'});
  const maxpool1 = builder.maxPool2d(
      conv1, {windowDimensions: [3, 3], strides, layout});
  const fire2 = await buildFire(maxpool1, 'fire2');
  const fire3 = await buildFire(fire2, 'fire3');
  const fire4 = await buildFire(fire3, 'fire4');
  const maxpool4 = builder.maxPool2d(
      fire4, {windowDimensions: [3, 3], strides, layout});
  const fire5 = await buildFire(maxpool4, 'fire5');
  const fire6 = await buildFire(fire5, 'fire6');
  const fire7 = await buildFire(fire6, 'fire7');
  const fire8 = await buildFire(fire7, 'fire8');
  const maxpool8 = builder.maxPool2d(
      fire8, {windowDimensions: [3, 3], strides, layout});
  const fire9 = await buildFire(maxpool8, 'fire9');
  const conv10 = await buildConv(fire9, 'conv10');
  const averagePool2d = builder.averagePool2d(
      conv10, {windowDimensions: [13, 13], layout});
  const reshape = builder.reshape(averagePool2d, [1, null]);
  return builder.softmax(reshape);
}

const build = async (outputOperand) => {
  graph = await builder.build({'output': outputOperand});
}

const compute = async (inputBuffer, outputBuffer) => {
  const inputs = {'input': inputBuffer};
  const outputs = {'output': outputBuffer};
  await context.compute(graph, inputs, outputs);
};

promise_setup(async () => {
  context = await navigator.ml.createContext();
  builder = new MLGraphBuilder(context);
}, {explicit_timeout: true, test_timeout: 180000});

promise_test(async () => {
  const inputElement = await createInputElement('img', './test.jpg');
  const inputBuffer = getInputTensor(inputElement, {
      mean: [127.5, 127.5, 127.5],
      std: [127.5, 127.5, 127.5],
      inputLayout: 'nhwc',
      inputDimensions: [1, 224, 224, 3],
  });
  const outputBuffer = new Float32Array(sizeOfShape(outputDimensions));
  console.time('[total] Load + build graph Squeenet 1.0 NHWC model');
  const outputOperand = await load();
  graph = await builder.build({'output': outputOperand});
  console.timeEnd('[total] Load + build graph Squeenet 1.0 NHWC model');
  const inputs = {'input': inputBuffer};
  const outputs = {'output': outputBuffer};
  console.time('Compute with Squeenet 1.0 NHWC model');
  const results = await context.compute(graph, inputs, outputs);
  console.timeEnd('Compute with Squeenet 1.0 NHWC model');
  const labels = await fetchLabels('./labels1001.txt');
  const labelClasses = getTopClasses(results.outputs.output, labels);

  assert_equals(labelClasses[0].label, 'lesser panda', 'Get top 1 label OK.');
  assert_equals(labelClasses[1].label, 'teddy', 'Get top 2 label OK.');
  assert_equals(labelClasses[2].label, 'giant panda', 'Get top 3 label OK.');
}, 'test SqueezeNet 1.0 model with NHWC layout');
</script>
